Create new component for updating tri
